Move more to abstact newick controller.
