+++ src/main/java/nl/tudelft/context/controller/NewickController.java
@@ -1,21 +1,14 @@
 package nl.tudelft.context.controller;
 
-import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.ReadOnlyObjectProperty;
-import javafx.beans.property.SimpleObjectProperty;
-import javafx.fxml.FXML;
-import javafx.scene.Group;
 import javafx.scene.control.Label;
 import javafx.scene.control.MenuItem;
-import javafx.scene.control.ProgressIndicator;
 import nl.tudelft.context.drawable.DrawableEdge;
 import nl.tudelft.context.drawable.NewickLabel;
 import nl.tudelft.context.model.newick.Newick;
 import nl.tudelft.context.model.newick.selection.None;
 
-import java.net.URL;
 import java.util.List;
-import java.util.ResourceBundle;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -27,26 +20,9 @@ import java.util.stream.Collectors;
 public final class NewickController extends DefaultNewickController {
 
     /**
-     * ProgressIndicator to show when the tree is loading.
-     */
-    @FXML
-    ProgressIndicator progressIndicator;
-
-    /**
-     * The container of the newick tree.
-     */
-    @FXML
-    Group newick;
-
-    /**
      * The main controller used to set views.
      */
     MainController mainController;
-
-    /**
-     * The newick object, can change.
-     */
-    ObjectProperty<Newick> newickObjectProperty;
 
     /**
      * The current selection of the tree.
@@ -59,11 +35,6 @@ public final class NewickController extends DefaultNewickController {
     GraphController graphController;
 
     /**
-     * Property with Newick tree.
-     */
-    ReadOnlyObjectProperty<Newick> newickIn;
-
-    /**
      * Init a controller at newick.fxml.
      *
      * @param mainController MainController for the application
@@ -72,42 +43,14 @@ public final class NewickController extends DefaultNewickController {
     public NewickController(final MainController mainController,
                             final ReadOnlyObjectProperty<Newick> newickIn) {
 
-        this.mainController = mainController;
+        super(newickIn);
 
-        this.newickIn = newickIn;
+        this.mainController = mainController;
 
         loadFXML("/application/newick.fxml");
     }
 
-    /**
-     * Called to initialize a controller after its root element has been
-     * completely processed.
-     *
-     * @param location  The location used to resolve relative paths for the root object, or
-     *                  <tt>null</tt> if the location is not known.
-     * @param resources The resources used to localize the root object, or <tt>null</tt> if
-     *                  the root object was not localized.
-     */
     @Override
-    public void initialize(final URL location, final ResourceBundle resources) {
-
-        newickObjectProperty = new SimpleObjectProperty<>();
-
-        newickObjectProperty.addListener((observable, oldValue, newValue) -> {
-            showTree(newValue);
-        });
-
-        newickObjectProperty.bind(newickIn);
-
-        progressIndicator.visibleProperty().bind(newickObjectProperty.isNull());
-
-    }
-
-    /**
-     * Show the phylogenetic tree.
-     *
-     * @param newick newick to show
-     */
     protected void showTree(final Newick newick) {
         // Bind edges
         List<DrawableEdge> edgeList = newick.edgeSet().stream()
